Task 2.3.1 is COMPLETE

+++ src/main/java/web/controller/UsersController.java
@@ -28,7 +28,7 @@ public class UsersController {
     }
 
     @GetMapping("/{id}")
-    public String show(@PathVariable("id") int id, Model model){
+    public String show(@PathVariable("id") long id, Model model){
         model.addAttribute("user", userService.show(id));
         return "show";
     }
@@ -45,7 +45,7 @@ public class UsersController {
     }
 
     @GetMapping("/{id}/edit")
-    public String edit(Model model, @PathVariable("id") int id) {
+    public String edit(Model model, @PathVariable("id") long id) {
         model.addAttribute("user", userService.show(id));
         return "edit";
     }
@@ -58,7 +58,7 @@ public class UsersController {
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable("id") int id) {
+    public String delete(@PathVariable("id") long id) {
         userService.delete(id);
         return "redirect:/user";
     }
